+++ tests/testthat/test_MESSAGES.R
@@ -60,6 +60,10 @@ test_that("RFBCoptions() error messages",{
                "value out of range")
   expect_error(RFBCoptions(cex.ann=11),
                "value out of range")
+  expect_error(RFBCoptions(cex.scaleBar=0),
+               "value out of range")
+  expect_error(RFBCoptions(cex.scaleBar=11),
+               "value out of range")
 })
 
 test_that("iGetopt() error messages",{